Revamped the whole irrigation loop
It tries to prevent missing events by a tight polling loop when close to
an event. It will also handle events that may be missed in case other
tasks get in the way or the previous event processing takes too long.

At the end the loop tries to be smart on how long to sleep to keep the
jitter low. This is done by measuring the loop time itself. It uses OS
ticks to be more accurate than the TimeSystem (1sec accuracy).