Completely reimplemented SNTP handling
SNTP request are now fully controlled by the IrrigationController. This
is due to the fact that a time update could cause an active irrigation
to be canceled. Therefore the IrrigationController plans SNTP updates if
irrigation is idle.

Also, a patch of the lwIP SNTP code is needed so that hooks to setting
the new time can be inserted. This is needed to signalize SNTP time
updates to the IrrigationController.